+++ trunk/src/decode/ltrnet/LTRCallEvent.java
@@ -25,14 +25,14 @@ import decode.DecoderType;
 public class LTRCallEvent extends CallEvent
 {
 	private int mChannel;
-	private int mFrequency;
+	private long mFrequency;
 	
 	public LTRCallEvent( CallEventType callEventType, 
 							 AliasList aliasList,
 							 String fromID,
 							 String toID,
 							 int channel,
-							 int frequency,
+							 long frequency,
 							 String details )
     {
 	    super( DecoderType.LTR_STANDARD, callEventType, 
@@ -101,7 +101,7 @@ public class LTRCallEvent extends CallEvent
     }
 
 	@Override
-    public int getFrequency()
+    public long getFrequency()
     {
 	    return mFrequency;
     }
@@ -117,7 +117,7 @@ public class LTRCallEvent extends CallEvent
 		private String mToID;
 		private String mDetails;
 		private int mChannel;
-		private int mFrequency;
+		private long mFrequency;
 
 		public Builder( CallEventType callEventType )
 		{
@@ -148,7 +148,7 @@ public class LTRCallEvent extends CallEvent
 			return this;
 		}
 		
-		public Builder frequency( int frequency )
+		public Builder frequency( long frequency )
 		{
 			mFrequency = frequency;
 			return this;
